+++ library/src/main/java/org/mustangproject/Charge.java
@@ -1,5 +1,6 @@
 package org.mustangproject;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.mustangproject.ZUGFeRD.IAbsoluteValueProvider;
 import org.mustangproject.ZUGFeRD.IZUGFeRDAllowanceCharge;
 
@@ -151,6 +152,7 @@ public class Charge implements IZUGFeRDAllowanceCharge {
 	 * @return true since it is supposed to be calculated negatively
 	 */
 	@Override
+	@JsonIgnore
 	public boolean isCharge() {
 		return true;
 	}