+++ library/src/test/java/org/mustangproject/ZUGFeRD/DeSerializationTest.java
@@ -49,4 +49,86 @@ public class DeSerializationTest extends TestCase {
 		assertEquals("info@company.com", fromJSON.getSender().getUriUniversalCommunicationID());
 
 	}
+
+	public void testAllowanceRead() throws JsonProcessingException {
+
+		ObjectMapper mapper = new ObjectMapper();
+
+		// [{"stringValue":"a","intValue":1,"booleanValue":true},
+		// {"stringValue":"bc","intValue":3,"booleanValue":false}]
+
+		Invoice fromJSON = mapper.readValue("{\n" +
+			"  \"documentCode\": \"380\",\n" +
+			"  \"number\": \"471102\",\n" +
+			"  \"ownOrganisationName\": \"Lieferant GmbH\",\n" +
+			"  \"currency\": \"EUR\",\n" +
+			"  \"issueDate\": \"2018-03-03T23:00:00.000+00:00\",\n" +
+			"  \"deliveryDate\": \"2018-03-03T23:00:00.000+00:00\",\n" +
+			"  \"sender\": {\n" +
+			"    \"name\": \"Lieferant GmbH\",\n" +
+			"    \"zip\": \"80333\",\n" +
+			"    \"street\": \"Lieferantenstraße 20\",\n" +
+			"    \"location\": \"München\",\n" +
+			"    \"country\": \"DE\",\n" +
+			"    \"taxID\": \"201/113/40209\",\n" +
+			"    \"vatID\": \"DE123456789\",\n" +
+			"    \"vatid\": \"DE123456789\"\n" +
+			"  },\n" +
+			"  \"recipient\": {\n" +
+			"    \"name\": \"Kunden AG Mitte\",\n" +
+			"    \"zip\": \"69876\",\n" +
+			"    \"street\": \"Kundenstraße 15\",\n" +
+			"    \"location\": \"Frankfurt\",\n" +
+			"    \"country\": \"DE\"\n" +
+			"  },\n" +
+			"  \"grandTotal\": 234.43,\n" +
+			"  \"zfitems\": [\n" +
+			"    {\n" +
+			"      \"price\": 9.9,\n" +
+			"      \"quantity\": 20,\n" +
+			"      \"tax\": null,\n" +
+			"      \"grossPrice\": null,\n" +
+			"      \"lineTotalAmount\": null,\n" +
+			"      \"basisQuantity\": 1,\n" +
+			"      \"detailedDeliveryPeriodFrom\": null,\n" +
+			"      \"detailedDeliveryPeriodTo\": null,\n" +
+			"      \"id\": null,\n" +
+			"      \"product\": {\n" +
+			"        \"unit\": \"H87\",\n" +
+			"        \"name\": \"Trennblätter A4\",\n" +
+			"        \"taxCategoryCode\": \"S\",\n" +
+			"        \"attributes\": null,\n" +
+			"        \"vatpercent\": 19\n" +
+			"      },\n" +
+			"      \"value\": 9.9\n" +
+			"    }\n" +
+			"  ],\n" +
+			"  \"tradeSettlement\": null,\n" +
+			"  \"ownTaxID\": \"201/113/40209\",\n" +
+			"  \"ownVATID\": \"DE123456789\",\n" +
+			"  \"ownStreet\": \"Lieferantenstraße 20\",\n" +
+			"  \"ownZIP\": \"80333\",\n" +
+			"  \"ownLocation\": \"München\",\n" +
+			"  \"ownCountry\": \"DE\",\n" +
+			"  \"zfallowances\": [\n" +
+			"    {\n" +
+			"      \"totalAmount\": 1,\n" +
+			"      \"taxPercent\": 19,\n" +
+			"      \"reason\": \"Sondernachlass\",\n" +
+			"      \"reasonCode\": null,\n" +
+			"      \"categoryCode\": \"S\",\n" +
+			"      \"charge\": false\n" +
+			"    }\n" +
+			"  ]\n" +
+			"}", Invoice.class);
+		TransactionCalculator tc=new TransactionCalculator(fromJSON);
+		assertEquals(tc.getGrandTotal(),new BigDecimal("234.43"));
+		assertEquals(fromJSON.getNumber(), fromJSON.getNumber());
+		assertEquals(fromJSON.getZFItems().length, fromJSON.getZFItems().length);
+
+	}
+
+
+
+
 }